requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
statsmodels>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
geopandas>=0.14.0
requests>=2.31.0
tqdm>=4.66.0
openai>=1.0.0
anthropic>=0.20.0
python-dotenv>=1.0.0
pycountry>=22.3.5
esda>=2.5.0
libpysal>=4.8.0
